Moved some files from the Private folder to the Public folder
