+++ Source/GitSourceControl/Public/GitSourceControlCommand.h
