+++ Source/GitSourceControl/Public/GitSourceControlModule.h
