+++ Source/GitSourceControl/Public/GitSourceControlProvider.h
