+++ Source/GitSourceControl/Public/GitSourceControlSettings.h
