+++ Source/GitSourceControl/Public/GitSourceControlUtils.h
