+++ Source/GitSourceControl/Public/IGitSourceControlWorker.h
